Add SetOf, FiniteSet, and some simple unit tests.
SetOf is a class for (possibly infinite) sets of elements. For our
purposes here, the most important parts of the class are 'univ',
'none', 'intersection', and 'sameAs'.

I've also added FiniteSet, which is just a newtype of Data.Set.
When the element types for 'e' is Bounded and Enum in addition to Ord,
'FiniteSet e' is an instance of 'SetOf e', since we can define
the 'univ' set.

The unit tests check the 'SetOf' test properties on 'FiniteSet Five',
where Five is a five-element enumeration.