Added internal zero-, one-, and two-variable constraints.
This adds three simple internal constraint types for
separately representing zero-variable, one-variable,
and two-variable input constraints, along with a function
for taking a list of input constraints and transforming
them into a triple of lists of the three internal
constraints.

As it stands, this code is hard to unit test because it
is not exported from the Solver module. I think I'll
split this out into a hidden internal module, add a
function for converting the triple of lists back
into a list of input constraints, and then add a
unit test to make sure this code is sane.