Move the NoVars, OneVar, TwoVars constraints into their own module.
In order to make it straightforward to unit test the various parts
of the solver, I've decided to move NoVars, OneVar, TwoVars and
the 'splitConstraints' function into their own module,
BRC.Solver.ZeroOneTwo. I may make the module hidden within the
cabal package at some point, but for now I'm leaving it visible.

I'd like to write a unit test for 'splitConstraints'; to do
that, I'll need to add functions for translating NoVars, OneVar,
and TwoVars back into the input Constraint. I'll do that in
the next commit.