Add a unit test for converting Constraints to and from internal form.
In order to test what was called 'splitConstraints', I added a
function to reassemble the input Constraints from the lists
of NoVars, OneVar, and TwoVars. This allows us to test for
arbitrary lists of input constraints that we can convert to
internal form and back.

In the process, I had to use phantom types in NoVars and TwoVars
in order to avoid type-checking ambiguities:

 NoVars e    is now   NoVars v e
 TwoVars v   is now   TwoVars v e

even though NoVars doesn't use type 'v' and TwoVars doesn't
use type 'e'.

The new test is in tests/ZeroOneTwoProps. In the process,
I added a generator for arbitrary test variables in
tests/Variables.