Add the solver monad, and its associated state.
The solver monad is both an error monad and a state monad. The state
comprises the lists of zero-variable, one-variable, and two-variable
internal constraints derived from the input constraints, along with
a running record of possible assignments to the variables.

All variables start out unconstrained. The solver will run in stages
that progressively refine an estimate of the possible bindings
for each variable. In the final stage, we generate the Cartesian
product of these possible bindings, filtering out those bindings
that violate any of the two-variable constraints.

Of course, I still have to write the stages. I'll do that in
the next few commits.