Correct the signature of Solver.solve.
Since the solver now runs in an error monad, the 'solve'
function should return either an error message or a list
of lists of bindings.

I'm mulling whether there is really any advantage in having
a Binding type instead of just ordered pairs. It will be a
trivial simplification if we remove Binding, so I'll defer
that until later.

If we do keep Binding, it might be a good idea to have an
Assignment type too.