Implement stage 1 of the solver, and add the 'smallints' example.
Stage 1 searches for contradictions in the zero-variable
constraints. I could have written a unit test for this, but
I decided that an example would be a good idea, so I've
added the 'smallints' directory.

The advantage of the 'smallints' example is that it allows
you to try out simple toy examples without having to type
a lot or write a fancy generator for satisfiable sets of
constraints. If somebody wants to try the solver out,
it would be an easy way to do it.

I've confirmed for the 'smallints' example that when I
put down a false constraint like

 10 <= 9

I get an error message, and when the constraints are
all true there is no error message. Currently, the
error messages are gross; I do need to put in a
configurable formatter for the constraints. But that
can wait for a bit.