Implement stage 2 of the solver, one-variable constraint handling.
The implementation of 'applyOneVarConstraints' is a little
bit nasty because of an attempt to generate a good error
message, modulo problems with the formatting of
Constraint itself. (Fixing that really requires a
configurable formatter, something best left to a separate
commit.)

To generate decent messages, we maintain a mapping
from variables to the constraints that have been
applied to them. If, at a particular constraint 'c',
we discover that we've now got the empty set, we'll
list the constraints that we've applied to that
variable, since the contradiction must be there.