Minor refactoring in 'Solver.applyOneVarConstraints'.
It's a bit weird to not immediately update the history after
applying the update on the possible assignments for a variable.
Now we update the history first, and then don't have to
remember to prepend the constraint we just applied. I forgot
to do that at first and the resulting message is an outright
lie, since it lists only a set of constraints that was not
contradictory.

I'm starting to think that a 'Stage2Monad' might have been
a good way to pass the additional state around and perhaps
thereby simplify 'applyOneVarConstraints'. I might experiment
with that for stage 3.