Added the BRC.Solver.Assignments module, and an update in Idea.txt.
It took a couple of days thinking it through, but this is a simple
but quite generic module for producing a Cartesian product subject
to pair-wise constraints on variables. It's certainly general
enough to do what we need in the BRC solver.

I've updated Idea.txt but it's still quite rough. I think the
code might be easier to understand, which is probably a good thing.