Avoid non-terminating searches in 'assignments'.
It's easy to to naively get a non-terminating search. For example,
giving 'subjectTo' predicates that can be unsatisfiable on unbounded
lists is a good way. Consider:

 subjectTo [(("x","y"), (>))] [("x", [1,2]), ("y" [1..])]

There are obviously satisfying assignments, like

 [("x", 2), ("y",1)]

However, we'd never find them. When "x" is assigned the trial value 1,
the possible values for "y" are

 filter (1 >) [1..]

and any attempt to test for and empty list or extract the head will
fail to terminate, since we've applied an unsatisfiable predicate to
an unbound list.

Now, 'subjectTo' could take an argument that limits the initial lists
of possible values. However, the problem is more general than that. In
'assignments', if the 'nextVar' function produces

 NextVar v x restX u

and 'restX' is an unbounded list, then for problems with no solution,
the search may be non-terminating.  If the first variable has an
unbounded list of trial values, then we will never finish proving that
they are all dead-ends.

To solve this, I've added a SearchOptions type with a setting for the
maximum number of trials when a variable becomes bound. In
'pushBinding', this is used to limit the actual list of remaining
trial values. In the filter functions built by 'subjectTo', it's also
used to limit the initial lists of possible values.