Fix a bug in the 'nextVar' function in 'subjectTo'.
The nextVar function is supposed to filter out possible values
for X using any predicate for (X,X). However, the code was
written like this:

   if null (filterSelf v es)
   then DeadEnd
   else NextVar v (head es) (tail es) rest

The 'else' part is wrong, because 'filterSelf' may eliminate
only some of the values in 'es', but we use *all* the values
in 'es' to build the NextVar. We should be using only the
values that pass the filter. Now it looks like:

   let es' = filterSelf v es
   in if null es'
      then DeadEnd
      else NextVar v (head es') (tail es') rest

I found this with a QuickCheck unit test I've got in progress,
solving constraints on Bool-valued variables. QuickCheck is
so very awesome.