First unit test for BRC.Solver.Assignments.
It took a while to build up the code for this, but I've
got a test for finding unique solutions, and some supporting
stuff for some other tests where the solutions need not
be unique.

This test takes a generated assignment on bounded, enumerable,
ordered domain, and then builds a set of less-than constraints
that have a unique solution. If X is assigned V, then

- when V is the domain minimum, we add constraint

  X < succ V

- when V is the domain maximum, we add constraint

  pred V < X

- and otherwise, we add the two constraints

  pred V < X   and
  X < succ V

This uniquely determines X, and we can convert the constraints
into predicates acceptable to 'subjectTo'. Since the constraints
have a unique solution,

1. 'subjectTo' should find exactly one assignment, provided
   the maxTrialsPerVariable is set to the size of the domain

2. the one assignment must be the same as the original
   assignment from which we derived the constraints

While doing this, I found a bug in the 'nextVar' function
in 'subjectTo' (addressed in the last commit), and I also
found a shortcoming in the interface of that function:
it silently discards all but one predicate on the same
pair of variables. For example, suppose the domain is
{1,2,3,4,5}, and X is assigned 3 in the test assignment.
In the constraints we derive from the assignment, we have
both

  2 < X    and    X < 4

This translates into the 'subjectTo' predicates

 ((X,X), const (2<))  and   ((X,X), const (<4))

But 'subjectTo' assumes that each predicate has
a unique pair of variables in the first component.
So if you give 'subjectTo' the predicate list

 [((X,X), const (2<)), ((X,X), const (<4))]

it will silently discard one of the predicates. It's up
to the client code to conjoin the predicates into one,
like so

 [((X,X), (\x y -> const (2<) x y && const (<4) x y)]

This is a definite interface wart, because it would be
easy for 'subjectTo' to conjoin the predicates itself.

I'll likely fix that problem after I add the next test.