Add a test that 'subjectTo' finds all solutions to satisfiable constraints.
The test generates an assignment and then derives the two-variable
less than constraints. It compares the results of 'subjectTo'
to those of a brute-force solver.

The size of the problems has to be seriously constrained to avoid
a giant brute-force problem space. Right now I limit the size of
the seed assignment S so that

 length S = floor (logBase D 100000)

where D is the size of the domain. Right now I'm testing with
'Five' as the domain, which limits us to problems with no more
than seven variables.

In the first test, where there are unique solutions, I also
used domain Five, but I didn't have to constrain the size
of the assignments. That's encouraging -- at least in that
case I know 'assignments' will perform reasonably well.