Add a module for grouping variables by their 'influence' classes.
BRC.Solver.Influence exports a single method for computing
the 'influence' classes of variables, where 'infleunces' is
the reflexive-transitive closure of a simple 'constrains'
relation.

I'm using the FGL to do this. I didn't remember there being
such a hassle building a graph, but perhaps I didn't find
the right functions to do it. The translation between
variables and graph nodes -- ints -- is a little annoying,
though not difficult to do.