Remove the 'Binding' type in BRC.Solver.
This type really has no advantages over an ordinary ordered
pair, unless the variables and the values happen to have the
same type. It's very easy for client code to avoid that though,
and it will make it simpler to use 'assignments' from
BRC.Solver.Assignments to implement the last stage of the solver.