Add SolverOptions, and improve error message formatting.
Before implementing the final stage of the solver, where
we'll use the 'assignments' function to solve the two-variable
constraints, we need to add options for the solver in order
to pass the 'maxTrialsPerVariable' to 'assignments'.

I've added SolverOptions in module BRC.Solver.Options. This
includes the 'maxTrialsPerVariable', and also includes
formatting functions for variables, values, sets, and constraints.

The 'solver' function now takes a SolverOptions argument.
In BRC.Solver.Monad there are now convenience functions for
getting the max trials per variable and formatting variables,
values, sets, and anything that is an instance of ToConstraint.
'solver' itself now uses the formatting functions to produce
error messages, instead of using 'show'.

This has a couple of benefits:

1. The variable and value type no longer need be instances
   of Show just to run the solver. The output of Show for
   many types isn't suitable for user-facing messages, so
   insisting on using Show is a bad choice.

   There are default options that do use show, but now
   it's just that choice of options that requires instances
   of Show.

2. The formatting can be better and more flexible. I've used
   this to greatly improve the format of error messages produced
   by the 'smallints' example.