Give 'influenceClasses' a list of supplemental variables.
When starting to implement phase 3 of the solver, I realized
that 'influenceClasses' would not in general produce the
singleton classes, since it didn't know about variables that
were involved in only one-variable constraints. Rather
than compute the single-variable classes in the solver,
I've given an 'extraVars' parameter to 'influenceClasses'.

In the process, I realized that there was some use for a
type class

 class VariablesIn v t | t -> v where
   variablesIn :: t -> [v]

to make extracting variables a little more uniform. I've
added one in the BRC.Solver.VariablesIn module, and
slightly simplified BRC.Solver.Influence as a result.