Fix a silly bug in the smallints example.
The module 'SmallIntConstraints' in the 'smallints' example
has a silly bug in the definition of 'lteRel'. This represents
the less-than-or-equals relation, but 'lteRight' was defined
incorrectly. When

  x <= y

it was excluding 'x' from the set. Of course this is wrong.
'x' should be included in both 'leftOf lteRel x' and
'rightOf lteRel x'.