Implement 'enumerateAssignments' in BRC.Solver.
This commit implements the third phase of the solver,
'enumerateAssignments'. There is more code here than I'd
ordinarily like in a single commit, but making the
changes was reasonably straightforward. There is almost
enough here to make into a separate module.

'enumerateAssignments' computes the influence classes
for the variables occurring in the constraints, solves
each influence class independently, and then combines
the results. To solve an influence class, we use the
'assignments' function from BRC.Solver.Assignments.

'assignments' requires a 'nextVar' function and an 'effects'
function. I've implemented these using a new 'Unbounds'
type for capturing the state of the unbound variables.
Unbounds tracks the current sets of possible values by
variable, the constraints by the variables they involve,
and the binary relation used to compute the effect of
the constraints.

In the process, I discovered that the SolverState needs
to carry the binary relationship. I wish I had realized
before, because those changes would have been nice to
do separately.

At this point, there are a couple of things missing:

1. if there is a two-variable constraint that uses
   the same variable twice, we don't compute the
   consequences in NextVar. That's a real bug,
   since we could say that there are solutions for
   a problem that is actually unsatisfiable, or that
   there are more solutions than there really are.

2. we don't pick a variable with a minimum-size
   possible set size

I don't want to write unit tests for this until (1)
is fixed, because there is a good chance the tests will
fail. Fixing (2) is a performance optimization.