Fix serious bugs: unconstrained variables vanish, extra assignments.
In BRC.Solver.solveClass, I wasn't careful to ensure that
the mapping from class variables to sets

1. contained a mapping from 'v' to 'univ' for every
   otherwise unconstrained variable

2. contained mappings for only the class variables

Violating (1) makes the solver think there are no
solutions to

 x <= y

The problem is that neither 'x' nor 'y' obtain a
set of potential bindings in the previous phases of
the solver, since there are no one-variable constraints
on either. The solver then immediately thinks its
done, because the map from variables to sets is empty.

Violating condition (2) can cause the solver to emit
bindings for variables more than once, because the
map from variables to sets can contain mappings for
variables that are not in the given equivalence class.
However, the 'nextVar' function assumes that every
entry in the map is fair game.

I've added two additional example problems in 'smallints'.
When doing this, I realized that I didn't even test an
example involving variables with no single-variable
constraints -- that's the cost of haste.